requirements.txt
torch==2.0.1+cu118
torchvision==0.15.2+cu118
hydra-core==1.3.2
omegaconf==2.3.0
ray[tune]>=2.7
torchmetrics>=1.4 